requirements.txt
Flask==2.3.2
tensorflow==2.13.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.1
Pillow==10.0.0
werkzeug==2.3.2
matplotlib===3.9.2
pickle
io
